Reorganize modules and simplify effect structure
- Move effects into their respective modules
- Rename Nostr.Effects.* to Nostr.*
- Combine IDGen and CurrentTime into Util effect
